InfluxdbAdaptor/requirements.txt
requests
paho-mqtt
influxdb
cherrypy